entendendo como criar meu prórpio hook a fim de facilitar a reutilizacao da funcao em outros pontos do código. Entendendo como o hook reducer pode simplificar lógicas extensas ao inves do useState que pode complicar criacao de metodos complexos. Endendi como utilizar o hook useMemo tendo a possibilidade de escolher os estados que vao ser atualizados de acorodo com o metodo, melhorando a performace do que aplicativo (uma vez que ele nao precisa recarregar todos os estados sempre que algum deles mudarem)
