+++ src/mocks/produtores.js
@@ -14,31 +14,31 @@ const produtores = {
     {
       nome: 'Green',
       imagem: green,
-      distancia: `${gerarNumeroRandomico(1, 500)}m`,
+      distancia: `${gerarNumeroRandomico(1, 500)}`,
       estrelas: gerarNumeroRandomico(1, 5),
     },
     {
       nome: 'Salad',
       imagem: salad,
-      distancia: `${gerarNumeroRandomico(1, 500)}m`,
+      distancia: `${gerarNumeroRandomico(1, 500)}`,
       estrelas: gerarNumeroRandomico(1, 5),
     },
     {
       nome: 'Jenny Jack',
       imagem: jennyJack,
-      distancia: `${gerarNumeroRandomico(1, 500)}m`,
+      distancia: `${gerarNumeroRandomico(1, 500)}`,
       estrelas: gerarNumeroRandomico(1, 5),
     },
     {
       nome: 'Grow',
       imagem: grow,
-      distancia: `${gerarNumeroRandomico(1, 500)}m`,
+      distancia: `${gerarNumeroRandomico(1, 500)}`,
       estrelas: gerarNumeroRandomico(1, 5),
     },
     {
       nome: 'Potager',
       imagem: potager,
-      distancia: `${gerarNumeroRandomico(1, 500)}m`,
+      distancia: `${gerarNumeroRandomico(1, 500)}`,
       estrelas: gerarNumeroRandomico(1, 5),
     },
   ],
