+++ src/telas/Home/componentes/Produtor.js
@@ -1,13 +1,22 @@
-import React, {useState} from 'react';
+import React, {useReducer, useMemo} from 'react';
 import {View, TouchableOpacity, Text, Image, StyleSheet} from 'react-native';
 import Estrelas from '../../../componentes/Estrelas';
 
+const distanciaMetros = distancia => {
+  return `${distancia}m`;
+};
+
 export default function Produtor({nome, imagem, distancia, estrelas}) {
-  const [selecionado, setSelecionado] = useState(false);
+  //const [selecionado, setSelecionado] = useState(false);
+  const [selecionado, inverterSelecionado] = useReducer(
+    selecionado => !selecionado,
+    false,
+  );
+
+  const distanciaTexto = useMemo(() => distanciaMetros(distancia), [distancia]);
+
   return (
-    <TouchableOpacity
-      style={estilos.cartao}
-      onPress={() => setSelecionado(!selecionado)}>
+    <TouchableOpacity style={estilos.cartao} onPress={inverterSelecionado}>
       <Image source={imagem} style={estilos.imagem} accessibilityLabel={nome} />
       <View style={estilos.info}>
         <View>
@@ -18,7 +27,7 @@ export default function Produtor({nome, imagem, distancia, estrelas}) {
             grande={selecionado}
           />
         </View>
-        <Text style={estilos.distancia}>{distancia}</Text>
+        <Text style={estilos.distancia}>{distanciaTexto}</Text>
       </View>
     </TouchableOpacity>
   );
