+++ src/telas/Home/componentes/Produtores.js
@@ -1,18 +1,11 @@
 import React, {useEffect, useState} from 'react';
 import {FlatList, StyleSheet, Text} from 'react-native';
 
-import {carregaProdutores} from '../../../servicos/carregaDados';
+import useProdutores from '../../../hooks/useProdutores';
 import Produtor from './Produtor';
 
 export default function Produtores() {
-  const [titulo, setTitulo] = useState('');
-  const [lista, setLista] = useState('');
-
-  useEffect(() => {
-    const retorno = carregaProdutores();
-    setTitulo(retorno.titulo);
-    setLista(retorno.lista);
-  }, []);
+  const [titulo, lista] = useProdutores();
 
   const TopoLista = () => {
     return <Text style={estilos.titulo}>{titulo}</Text>;
